# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_techinsight_be-techinsights-notification-service_internal_domain.Notification:
    properties:
      body:
        type: string
      created_at:
        type: string
      event_type:
        type: string
      extra:
        additionalProperties: true
        type: object
      id:
        type: string
      link:
        type: string
      read:
        type: boolean
      title:
        type: string
      user_id:
        type: string
    type: object
  github_com_techinsight_be-techinsights-notification-service_internal_domain.Preference:
    properties:
      in_app_enabled:
        type: boolean
      push_enabled:
        type: boolean
      user_id:
        type: string
    type: object
  github_com_techinsight_be-techinsights-notification-service_internal_domain.Subscriber:
    properties:
      device_token:
        type: string
      id:
        type: string
      platform:
        description: ios, android, web
        type: string
      user_id:
        type: string
      web_push_endpoint:
        type: string
      web_push_keys:
        additionalProperties:
          type: string
        type: object
    type: object
  github_com_techinsight_be-techinsights-notification-service_pkg_httputil.APIError:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Meta:
    properties:
      has_more:
        type: boolean
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.APIError'
        type: array
      meta:
        $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Meta'
    type: object
info:
  contact: {}
  description: REST API for the TechInsight notification service. Manages notifications,
    user preferences, device subscribers, and Centrifugo WebSocket tokens.
  title: TechInsight Notification Service API
  version: "1.0"
paths:
  /auth/centrifugo-token:
    post:
      consumes:
      - application/json
      description: Generate a JWT token for client WebSocket connection to Centrifugo
      parameters:
      - description: User ID
        in: body
        name: body
        required: true
        schema:
          properties:
            user_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  token:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
      summary: Generate Centrifugo connection token
      tags:
      - auth
  /notifications/{id}:
    delete:
      description: Delete a notification by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
      summary: Delete a notification
      tags:
      - notifications
    get:
      description: Get a single notification by its ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_internal_domain.Notification'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
      summary: Get a notification by ID
      tags:
      - notifications
  /notifications/{id}/read:
    patch:
      description: Mark a single notification as read by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
      summary: Mark a notification as read
      tags:
      - notifications
  /users/{userId}/notifications:
    get:
      description: Get paginated list of notifications for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter unread only
        in: query
        name: unread_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_internal_domain.Notification'
                  type: array
                meta:
                  $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Meta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
      summary: List user notifications
      tags:
      - notifications
  /users/{userId}/notifications/read-all:
    patch:
      description: Mark all unread notifications as read for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
            - properties:
                data:
                  properties:
                    updated:
                      format: int64
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
      summary: Mark all notifications as read
      tags:
      - notifications
  /users/{userId}/notifications/unread-count:
    get:
      description: Get the count of unread notifications for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
            - properties:
                data:
                  properties:
                    count:
                      format: int64
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
      summary: Count unread notifications
      tags:
      - notifications
  /users/{userId}/preferences:
    get:
      description: Get notification preferences for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_internal_domain.Preference'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
      summary: Get user preferences
      tags:
      - preferences
    put:
      consumes:
      - application/json
      description: Update notification preferences for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Preference settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_internal_domain.Preference'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
      summary: Update user preferences
      tags:
      - preferences
  /users/{userId}/subscribers:
    post:
      consumes:
      - application/json
      description: Register a device for push notifications (FCM/Web Push)
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Subscriber details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_internal_domain.Subscriber'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
      summary: Register a device subscriber
      tags:
      - subscribers
  /users/{userId}/subscribers/{token}:
    delete:
      description: Unregister a device from push notifications
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Device token
        in: path
        name: token
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_techinsight_be-techinsights-notification-service_pkg_httputil.Response'
      summary: Delete a device subscriber
      tags:
      - subscribers
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"